Change button color scheme in Login and Register components
The code diff above shows changes made to the Login.tsx and Register.tsx files in the auth/components directory. In the Login component, the colorScheme prop of the Button component was changed from 'green' to 'black', and the color prop was changed from the default (which would be 'inherit') to 'white'. Similarly, in the Register component, the color prop of the Heading component was changed from 'black' to 'white'. These changes were made to adjust the color scheme of the components.